fix: validation of kakao user id, token

+++ src/main/java/jupgo/jupgoserver/dto/user/SaveUserRequestDto.java
@@ -8,7 +8,7 @@ public class SaveUserRequestDto {
     private String nickname;
 
     private String email;
-    private Integer kakaoId;
+    private Long     kakaoId;
 
     private String code;
